+++ client/src/components/DiagnosticContainer.js
@@ -1,11 +1,11 @@
 import React from 'react';
-import { loadObservations } from '../../redux/actions/observations';
+import { loadObservations } from '../redux/actions/observations';
 import { useDispatch } from 'react-redux';
 import HealthLogToggle from './HealthLogToggle';
-import Scroll from '../Scroll'
+import Scroll from './Scroll'
 import Container from '@material-ui/core/Container';
-import Disclaimer from '../Disclaimer';
-import Subscribe from '../Subscribe';
+import Disclaimer from './Disclaimer';
+import Subscribe from './Subscribe';
 
 const dateOptions = {
   weekday: 'long',
